+++ src/utils/isIE/index.ts
@@ -1,0 +1,3 @@
+import isIE from './isIE';
+
+export default isIE;